+++ src/contracts/abi/ERC1271.json
